Server aliases now work with default aliases
You can now use server-specific aliases when scheduling game nights. Default aliases take priority over server aliases, so be careful when trying to add one of the same name. The first appearance of the alias used is taken, so make sure aliases do not conflict with others in the sheet.